move prepare_crr_df to component wrpapped function
